requirements.txt
django_import_export==2.2.0
Django==3.0.4
